Compute bottom azimut

Compute bottom azimut

Usage

get_bottom_azimut(aspect, north2x)

Arguments

aspect

double - Angle of slope bottom on the compass from the North, clockwise rotation (in degrees) northern aspect : 0, eastern aspect : 90, southern aspect : 180, western aspect : 270

north2x

double - Angle from North to x axis clockwise. (in degrees) Default correspond to a Y axis oriented toward the North.