Compute samsara ligth radiative balance

Compute samsara ligth radiative balance

Usage

sl_run(
  trees,
  monthly_rad,
  sensors = NULL,
  sensors_only = FALSE,
  latitude = 46,
  start_day = 1,
  end_day = 365,
  slope = 0,
  north_to_x_cw = 90,
  aspect = 0,
  cell_size = 10,
  n_cells_x = 10,
  n_cells_y = 10,
  soc = TRUE,
  height_anglemin = 10,
  direct_startoffset = 0,
  direct_anglestep = 5,
  diffuse_anglestep = 15,
  use_torus = TRUE,
  turbid_medium = TRUE,
  trunk_interception = TRUE,
  detailed_output = FALSE
)

Arguments

trees

a data.frame with one row for each tree and 14 columns describing tree composing the stand:

id: Unique id of the tree

x: X position of the tree within the stand

y: Y position of the tree within the stand

dbh_cm: Diameter at breast height (1.3m) of the tree trunk (in cm)

crown_type:

"E": symetric ellispoidal crown with the maximum crown radius at the middle of the crown depth. Crown radius is the mean value of the radius towards the fourth cardinal points

"P": symetric paraboloidal crown with the maximum crown radius at the crown base. crown radius is the mean value of the radius towards the fourth cardinal points

"2E": ellipsoidal crown composed of two semi-ellipsoid (one above and one below). Same as "E" but height of each semi-ellipsoid is given by hmax_m variable

"8E": ellispoidal crown composed of eight eighth of ellipsoids (four above and four below). Same as "2E" but radius of each eight of ellipsoid is given by rn_m, rs_m, re_m, rw_m variables.

"4P": paraboloidal crown composed of four fourth of paraboloid. Maximum crown radius at the crown base and radius of each fourth of paraboloid is given by rn_m, rs_m, re_m, rw_m variables.

h_m: Height of the tree (in meters)

hbase_m: Height of the base of the tree crown (in meters)

hmax_m: Height of the maximum radius of the tree crown (in meters)

rn_m: Radius of the tree crown toward Northern direction (in meters)

rs_m: Radius of the tree crown toward Southern direction (in meters)

re_m: Radius of the tree crown toward Eastern direction (in meters)

rw_m: Radius of the tree crown toward Western direction (in meters)

crown_openess: Crown openess of the crown (no unit) when considering porous envelop (i.e. constant proportion of light energy interception when a ray intercepts a crown)

crown_lad: Leaf Area Density (in m2 of leaves per m3 of crown) when considering turbid medium (i.e. density parameter of Beer Lambert law when a ray intercepts a crown)

monthly_rad

data.frame - Monthly horizontal radiation (Hrad) and diffuse to global ratio (DGratio) Computed with function samsaRaLight::sl_get_monthlyrad()

sensors

data.frame - Position of the sensor within the stand. Can be set to NULL if no sensor.

id: Unique id of the sensor

x: X position of the sensor within the stand

y: Y position of the sensor within the stand

h_m: Height above ground of the sensor (in meters)

sensors_only

boolean - To compute interception only for the sensors ?

latitude

double - Latitude of the plot (in degrees)

start_day

integer between 1 and 365 - First day of the vegetative period

end_day

integer between 1 and 365 - Last day of the vegetative period

slope

double - Slope of the plot (in degrees)

north_to_x_cw

double - Angle from North to x axis clockwise. (in degrees) Default correspond to a Y axis oriented toward the North.

aspect

double - Angle of slope bottom on the compass from the North, clockwise rotation (in degrees) northern aspect : 0, eastern aspect : 90, southern aspect : 180, western aspect : 270

cell_size

double - Length of the side of a squared cell composing the stand (in meters)

n_cells_x

integer - Number of cells of the X-axis side of the rectangle stand

n_cells_y

integer - Number of cells of the Y-axis side of the rectangle stand

soc

boolean - Standard Overcast Sky, if false: Uniform Overcast Sky

height_anglemin

double - Angle minimum between beam and soil (in degrees)

direct_startoffset

double - Angle at which to start first direct ray (in degrees)

direct_anglestep

double - Hour angle between two direct beams (in degrees)

diffuse_anglestep

double - Hour angle between two diffuse beams (in degrees)

use_torus

if True, compute light competition using borders modelled with a torus system, otherwise,borders are open grasslands

turbid_medium

If TRUE, crown are considered as turbid medium, otherwise, considered as porous envelope

trunk_interception

Consider interception of rays by trunks

detailed_output

boolean - If TRUE, sensors/cells/trees outputs also contain diffuse/direct energies, and sensors/cells outputs contain energies both on the slope and on a horizontal plane. If FALSE, energy given is only the total energy (sum of diffuse and direct) on the slope for trees and cells and on a horizontal plane for sensors.