SamsaRaLight: Forest light interception and radiation simulation

SamsaRaLight is an R package for simulating forest stand light interception, energy balance and radiation transfer using spatially explicit tree crowns, sensors and grid cells.

Details

The package provides tools to:

Build forest stands with trees, cells and sensors

Run ray-tracing based light simulations

Compute tree, cell and sensor energy variables

Summarize and visualize simulation outputs

See also

Author

Maintainer: Natheo Beauchamp beauchamp.natheo@gmail.com (R package developper)

Authors:

Gauthier Ligot gligot@uliege.be (Java Capsis developper and algorithm improvment) [copyright holder]

Francois de Coligny coligny@cirad.fr (Java Capsis developper)

Maxime Jaunatre maxime.jaunatre@inrae.fr (R package support developper)

Benoit Courbaud benoit.courbaud@inrae.fr (Algorithm and theory creator) [copyright holder]